Moved feedback stuff into own files Fixed cors issue with ajax requests
