+++ mentea/templates/base.html
@@ -20,6 +20,10 @@
     <link rel="stylesheet" href="{% static 'css/local.css' %}" />
     {% comment %} This is where a page's custom CSS will go, or a link to the
     relevant stylesheet I guess. {% endcomment %} {% endblock %}
+
+    <script type="text/javascript">
+      window.CSRF_TOKEN = "{{ csrf_token }}";
+    </script>
   </head>
   <body>
     <div id="pagetop">
@@ -53,7 +57,7 @@ main
 {% comment %} For scripts at the bottom of the page. {% endcomment %}
 {% block bottomscripts %}
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
-
+  
     <script src="https://code.jquery.com/jquery-3.6.0.min.js"></script>
     <script
       src="https://cdnjs.cloudflare.com/ajax/libs/TableDnD/0.9.1/jquery.tablednd.js"
